Created monitoring role during server provisioning and added continuous deployment to CircleCI
the monitoring role installs and configures the TICK stack for monitoring
i.e telegraph, influxdb, chronograf and kapacitor. Promethius is also
included for monitoring the django application while running.
CD in circleCI is in testing and uses a playbook to update the files
in the server and restart the service